done with rounded corner qr

+++ routes/createQr/create.js
@@ -9,6 +9,8 @@ const templates = require('../../qr/templates')
 const router = express.Router()
 const utils = require('../../utils')
 
+const defaultQR = require('./defaultQR')
+
 router.post('/create/:type', async (req, res) => {
     const {type} = req.params
     const options = req.body
@@ -18,30 +20,35 @@ router.post('/create/:type', async (req, res) => {
             filteredOptions[item] = options[item]
         }
     })
+    if(type === 'default'){
+        await defaultQR(res, filteredOptions)
+        return // no need to execute below code
+    }
+
     try{
+        const img = await createQR(filteredOptions.data, {filteredOptions, primaryColor: '#000000', secondaryColor: '#ffffff'}, false)
+        const array = await qrToArray(img, 2)
+        await utils.deleteImage(img, true)
+        let filePath = ''
         switch(type){
-            case 'default': {
-                const img = await templates.defaultQR(filteredOptions.data, filteredOptions)
-                res.status(201).json(successResponse(201, {
-                    url: utils.getImageHttpPath(img.split('/').at(-1))
-                }))
+            case 'circular': {
+                filePath = await templates.circular(array, filteredOptions)
                 break
             }
-            case 'circular': {
-                const img = await createQR(filteredOptions.data, {filteredOptions, primaryColor: '#000000', secondaryColor: '#ffffff'}, false)
-                const array = await qrToArray(img, 2)
-                const filePath = await templates.circular(array, filteredOptions)
-                await utils.deleteImage(img, true)
-                res.status(201).json(successResponse(201, {
-                    url: utils.getImageHttpPath(filePath.split('/').at(-1))
-                }))
+            case 'edge-rounded': {
+                filePath = await templates.edgeRounded(array, filteredOptions)
                 break
             }
         }
+        res.status(201).json(successResponse(201, {
+            url: utils.getImageHttpPath(filePath.split('/').at(-1))
+        }))
+        return
     }
     catch(err){
         console.log(err)
         res.status(500).json(failResponse(500, ['Server is facing some issue.']))
+        return
     }
 })
